+++ packages/backend/src/exploreCompiler.ts
@@ -150,9 +150,9 @@ export const compileMetricSql = (
                 {},
             );
     }
+
     if (metric.round !== undefined)
         renderedSql = `ROUND(${renderedSql}, ${metric.round})`;
-
     return renderedSql;
 };