+++ packages/frontend/src/hooks/useEcharts.ts
@@ -10,13 +10,14 @@ import {
     friendlyName,
     getDimensions,
     getFieldLabel,
+    hashFieldReference,
     MetricType,
+    Series,
     TableCalculation,
 } from 'common';
 import { useMemo } from 'react';
 import { useVisualizationContext } from '../components/LightdashVisualization/VisualizationProvider';
 import { getDimensionFormatter } from '../utils/resultFormatter';
-import { getPivotedDimension } from './usePlottedData';
 
 const getLabelFromField = (
     explore: Explore,
@@ -65,7 +66,7 @@ const getAxisTypeFromField = (
     }
 };
 
-const getEchartsTooltipConfig = (type: CartesianChart['series'][0]['type']) =>
+const getEchartsTooltipConfig = (type: Series['type']) =>
     type === CartesianSeriesType.BAR
         ? {
               show: true,
@@ -83,7 +84,7 @@ const getEchartsTooltipConfig = (type: CartesianChart['series'][0]['type']) =>
           };
 
 export type EChartSeries = {
-    type: CartesianChart['series'][0]['type'];
+    type: Series['type'];
     connectNulls: boolean;
     name?: string;
     color?: string;
@@ -125,129 +126,120 @@ export const getEchartsSeries = (
     explore: Explore,
     tableCalculations: TableCalculation[],
     originalData: ApiQueryResults['rows'],
-    chartConfig: CartesianChart,
+    cartesianChart: CartesianChart,
     pivotKey: string | undefined,
 ): EChartSeries[] => {
     if (pivotKey) {
-        const uniquePivotValues: string[] = Array.from(
-            new Set(originalData.map((row) => row[pivotKey])),
-        );
-        return chartConfig.series.reduce<EChartSeries[]>(
-            (sum, { yField, xField, type, flipAxes, label }) => {
-                const xAxisDimension = {
-                    name: xField,
-                    displayName: getLabelFromField(
-                        explore,
-                        tableCalculations,
-                        xField,
-                    ),
-                };
-                const groupSeries = uniquePivotValues.map((rawValue) => {
-                    const pivotedDimension = getPivotedDimension(
-                        rawValue,
-                        yField,
-                    );
-                    const value = getFormatterValue(
-                        rawValue,
-                        pivotKey,
-                        getDimensions(explore),
-                    );
-                    return {
-                        label,
-                        type,
-                        connectNulls: true,
-                        encode: {
-                            x: flipAxes ? pivotedDimension : xField,
-                            y: flipAxes ? xField : pivotedDimension,
-                            tooltip:
-                                type === CartesianSeriesType.BAR
-                                    ? [pivotedDimension]
-                                    : [xField, pivotedDimension],
-                            seriesName: pivotedDimension,
-                        },
-                        dimensions: [
-                            xAxisDimension,
-                            {
-                                name: pivotedDimension,
-                                displayName:
-                                    chartConfig.series.length > 1
-                                        ? `[${value}] ${getLabelFromField(
-                                              explore,
-                                              tableCalculations,
-                                              yField,
-                                          )}`
-                                        : value,
-                            },
-                        ],
-                        tooltip: {
-                            valueFormatter: valueFormatter(
-                                xField,
-                                yField,
+        return (cartesianChart.eChartsConfig.series || []).map<EChartSeries>(
+            (series) => {
+                const { flipAxes } = cartesianChart.layout;
+                const xField = hashFieldReference(series.encode.xRef);
+                const yFieldHash = hashFieldReference(series.encode.yRef);
+                const pivotField = series.encode.yRef.pivotValues?.find(
+                    ({ field }) => field === pivotKey,
+                );
+
+                const value = getFormatterValue(
+                    pivotField?.value,
+                    pivotKey,
+                    getDimensions(explore),
+                );
+                return {
+                    ...series,
+                    connectNulls: true,
+                    encode: {
+                        x: flipAxes ? yFieldHash : xField,
+                        y: flipAxes ? xField : yFieldHash,
+                        tooltip:
+                            series.type === CartesianSeriesType.BAR
+                                ? [yFieldHash]
+                                : [xField, yFieldHash],
+                        seriesName: yFieldHash,
+                    },
+                    dimensions: [
+                        {
+                            name: xField,
+                            displayName: getLabelFromField(
                                 explore,
+                                tableCalculations,
+                                xField,
                             ),
                         },
-                    };
-                });
-
-                return [...sum, ...groupSeries];
+                        {
+                            name: yFieldHash,
+                            displayName:
+                                cartesianChart.eChartsConfig.series &&
+                                cartesianChart.eChartsConfig.series.length > 1
+                                    ? `[${value}] ${getLabelFromField(
+                                          explore,
+                                          tableCalculations,
+                                          series.encode.yRef.field,
+                                      )}`
+                                    : value,
+                        },
+                    ],
+                    tooltip: {
+                        valueFormatter: valueFormatter(
+                            xField,
+                            series.encode.yRef.field,
+                            explore,
+                        ),
+                    },
+                };
             },
-            [],
         );
     } else {
-        return chartConfig.series.reduce<EChartSeries[]>(
-            (sum, { yField, xField, type, flipAxes, name, color, label }) => {
-                return [
-                    ...sum,
-                    {
-                        type,
-                        label,
-                        connectNulls: true,
-                        name,
-                        color,
-                        encode: {
-                            x: flipAxes ? yField : xField,
-                            y: flipAxes ? xField : yField,
-                            tooltip:
-                                type === CartesianSeriesType.BAR
-                                    ? [yField]
-                                    : [xField, yField],
-                            seriesName: yField,
-                        },
-                        dimensions: [
-                            {
-                                name: xField,
-                                displayName: getLabelFromField(
-                                    explore,
-                                    tableCalculations,
-                                    xField,
-                                ),
-                            },
-                            {
-                                name: yField,
-                                displayName: getLabelFromField(
-                                    explore,
-                                    tableCalculations,
-                                    yField,
-                                ),
-                            },
-                        ],
-                        tooltip: {
-                            valueFormatter: valueFormatter(
-                                xField,
-                                yField,
+        return (cartesianChart.eChartsConfig.series || []).reduce<
+            EChartSeries[]
+        >((sum, series) => {
+            const { flipAxes } = cartesianChart.layout;
+            const xField = hashFieldReference(series.encode.xRef);
+            const yField = hashFieldReference(series.encode.yRef);
+            return [
+                ...sum,
+                {
+                    ...series,
+                    connectNulls: true,
+                    encode: {
+                        ...series.encode,
+                        x: flipAxes ? yField : xField,
+                        y: flipAxes ? xField : yField,
+                        tooltip:
+                            series.type === CartesianSeriesType.BAR
+                                ? [yField]
+                                : [xField, yField],
+                        seriesName: yField,
+                    },
+                    dimensions: [
+                        {
+                            name: xField,
+                            displayName: getLabelFromField(
                                 explore,
+                                tableCalculations,
+                                xField,
                             ),
                         },
+                        {
+                            name: yField,
+                            displayName: getLabelFromField(
+                                explore,
+                                tableCalculations,
+                                yField,
+                            ),
+                        },
+                    ],
+                    tooltip: {
+                        valueFormatter: valueFormatter(xField, yField, explore),
                     },
-                ];
-            },
-            [],
-        );
+                },
+            ];
+        }, []);
     }
 };
 
 const getEchartAxis = (
-    validConfig: CartesianChart | undefined,
+    layout: CartesianChart['layout'],
+    series: EChartSeries[],
     explore: Explore,
     xAxisField: string | undefined,
     yAxisField: string | undefined,
@@ -257,8 +249,8 @@ const getEchartAxis = (
 
     let xAxisType = defaultXAxisType;
     let yAxisType = defaultYAxisType;
-    if (validConfig?.series[0].type === CartesianSeriesType.BAR) {
-        if (validConfig?.series[0].flipAxes) {
+    if (series[0].type === CartesianSeriesType.BAR) {
+        if (layout.flipAxes) {
             xAxisType = defaultXAxisType;
             yAxisType =
                 defaultYAxisType === 'value' ? 'category' : defaultYAxisType;
@@ -274,7 +266,7 @@ const getEchartAxis = (
 
 const useEcharts = () => {
     const {
-        cartesianConfig: { validConfig },
+        cartesianConfig: { validCartesianConfig },
         explore,
         plotData,
         originalData,
@@ -283,7 +275,7 @@ const useEcharts = () => {
     } = useVisualizationContext();
 
     const series = useMemo(() => {
-        if (!explore || !validConfig || !resultsData) {
+        if (!explore || !validCartesianConfig || !resultsData) {
             return [];
         }
 
@@ -291,27 +283,39 @@ const useEcharts = () => {
             explore,
             resultsData.metricQuery.tableCalculations,
             originalData,
-            validConfig,
+            validCartesianConfig,
             pivotDimensions?.[0],
         );
-    }, [explore, validConfig, resultsData, pivotDimensions, originalData]);
+    }, [
+        explore,
+        validCartesianConfig,
+        resultsData,
+        pivotDimensions,
+        originalData,
+    ]);
 
-    if (!explore || series.length <= 0 || plotData.length <= 0) {
+    if (
+        !explore ||
+        series.length <= 0 ||
+        plotData.length <= 0 ||
+        !validCartesianConfig
+    ) {
         return undefined;
     }
 
-    const [xAxis] = validConfig?.xAxes || [];
-    const [yAxis] = validConfig?.yAxes || [];
+    const [xAxis] = validCartesianConfig.eChartsConfig?.xAxis || [];
+    const [yAxis] = validCartesianConfig.eChartsConfig?.yAxis || [];
 
-    const xAxisField = validConfig?.series[0].flipAxes
-        ? validConfig?.series[0].yField
-        : validConfig?.series[0].xField;
-    const yAxisField = validConfig?.series[0].flipAxes
-        ? validConfig?.series[0].xField
-        : validConfig?.series[0].yField;
+    const xAxisField = validCartesianConfig.layout?.flipAxes
+        ? validCartesianConfig.layout?.yField?.[0]
+        : validCartesianConfig.layout?.xField;
+    const yAxisField = validCartesianConfig.layout?.flipAxes
+        ? validCartesianConfig.layout?.xField
+        : validCartesianConfig.layout?.yField?.[0];
 
     const { xAxisType, yAxisType } = getEchartAxis(
-        validConfig,
+        validCartesianConfig.layout,
+        series,
         explore,
         xAxisField,
         yAxisField,